+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 ## 1.1.1 (not yet released)
 
-(nothing yet)
+- [issue #1] Use realpath to find deps and doc files relative to nodedoc
+  script.
 
 
 ## 1.1.0